updated scripts to auto upload to cosmos added version information to files
