+++ .monitorial/run.ps1
@@ -1,0 +1,1 @@
+python .\run.py
